% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer.df}
\alias{cancer.df}
\title{Prostate cancer patient data}
\format{
A data.frame with 475 rows and 12 columns:
\describe{
\item{age}{Age in years.}
\item{wt}{Weight in pounds}
\item{pf}{Patient activity.}
\item{hx}{Family history of cancer.}
\item{sbp}{Systolic blood pressure.}
\item{dbp}{Diastolic blood pressure.}
\item{ekg}{Electrocardiogram code.}
\item{hg}{Haemoglobin}
\item{sz}{Tumour size.}
\item{sg}{Tumour index. }
\item{ap}{Acid phosphatase.}
\item{bm}{Bone metastatses}
}
}
\source{
D.P. Byar and S.B. Green "The choice of treatment for cancer patients
  based on covariate information - application to prostate cancer", Bulletin
  du Cancer 1980: 67:477--490, reproduced in D.A. Andrews and A.M. Herzberg
  "Data: a collection of problems from many fields for the student and
  research worker" p.261--274 Springer series in statistics, Springer-Verlag.
  New York.
}
\usage{
cancer.df
}
\description{
Data on 475 prostate cancer patients
}
\keyword{datasets}
