% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.r
\name{count.unique}
\alias{count.unique}
\title{Count the number of unique items ion a vector x}
\usage{
count.unique(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the number of unique items in \code{x}.
}
\description{
Count the number of unique items ion a vector x
}
\examples{
x = c(1, 2, 3)
count.unique(x)

x = c(1, 1, 1, 2, 3)
count.unique(x)
}
\author{
Murray Jorgensen
}
