% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_organise.r
\name{data_organise}
\alias{data_organise}
\title{Prepare data for use with multimix}
\usage{
data_organise(
  dframe,
  numClusters,
  niter = 1000,
  cdep = NULL,
  lcdep = NULL,
  minpstar = 1e-09
)
}
\arguments{
\item{dframe}{a data frame containing the data set you wish to model.}

\item{numClusters}{the clusters you wish to fit.}

\item{niter}{the maximum number of steps to that the EM agorithm will run
before terminating.}

\item{cdep}{a list of multivariate normal cells.}

\item{lcdep}{a list of location cells.}

\item{minpstar}{Minimum denominator for appliction of Bayes Rule.}
}
\value{
An object of class \code{multimixSettings} which is a \code{list} 
with the following elements:
\itemize{
   \item{\code{cdep}}{}
   \item{\code{clink}}{}
   \item{\code{cprods}}{}
   \item{\code{cvals}}{}
   \item{\code{cvals2}}{}
   \item{\code{dframe}}{}
   \item{\code{discvar}}{}
   \item{\code{dlevs}}{}
   \item{\code{dlink}}{}
   \item{\code{dvals}}{}
   \item{\code{lc}}{}
   \item{\code{lcdep}}{}
   \item{\code{lcdisc}}{}
   \item{\code{lclink}}{}
   \item{\code{lcprods}}{}
   \item{\code{lcvals}}{}
   \item{\code{lcvals2}}{}
   \item{\code{ld}}{}
   \item{\code{ldlevs}}{}
   \item{\code{ldlink}}{}
   \item{\code{ldvals}}{}
   \item{\code{ldxc}}{}
   \item{\code{mc}}{}
   \item{\code{md}}{}
   \item{\code{minpstar}}{}
   \item{\code{n}}{}
   \item{\code{nIt}}{}
   \item{\code{oc}}{}
   \item{\code{olink}}{}
   \item{\code{op}}{}
   \item{\code{ovals}}{}
   \item{\code{ovals2}}{}
   \item{\code{qq}}{the number of clusters in the model.}
}
}
\description{
Prepare data for use with multimix
}
\examples{
data(cancer.df)
mmObj = data_organise(cancer.df)
}
