% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eStep.R
\name{eStep}
\alias{eStep}
\title{The E(xpectation) step}
\usage{
eStep(P, D)
}
\arguments{
\item{P}{an object of class \code{multimixParamList}--see 
\code{\link{initParamList}} for more information.}

\item{D}{an object of class \code{multimixSettings}---see
\code{\link{data_organise}} for more information.}
}
\value{
a \code{list} containing two elements: a \code{matrix} named 
\code{Z}---see \code{\link{mStep}} for more information, and a scalar
\code{llik} containing the current value of the log-likelihood.
}
\description{
The E(xpectation) step
}
