% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_mixed_effects_model_lite.R
\name{fit_model_lite}
\alias{fit_model_lite}
\title{fit_model_lite: Fit a lighter version of the mixed effects model using initiating parameters}
\usage{
fit_model_lite(
  df_long,
  nGH = 40,
  fixed_pars = list(),
  full_param_names = lite_param_names,
  default_init = default_init_example
)
}
\arguments{
\item{df_long}{data in long format. Columns are: Subject_ID, Time, Binary_outcome}

\item{nGH}{number of nodes for optimizer}

\item{fixed_pars}{named list of parameters to fix values (i.e. no optimization on)}

\item{default_init}{named list of initial parameters to try optimizing on}
}
\value{
An object of class \code{multi_mix_model}, which is a list containing at least:
\describe{
\item{df_long}{The original \code{df_long} data frame used for fitting.}
\item{est}{Named numeric vector of estimated parameters.}
\item{u_hat}{Estimated random effects of size \verb{[N, 2]}}
\item{logLik}{Numeric. Log-likelihood of the fitted model.}
}
The object is intended to be used with S3 methods such as \code{print()}, \code{summary()}, and \code{plot()}.
}
\description{
This is called by fit_model_with_retries_lite using different initiating parameters.
Contrary to the full model, this model only fits a single random effect and
uses a scaling coefficient to approximate the second random effect
}
