% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParamList.R
\name{initParamList}
\alias{initParamList}
\title{Initialise the parameter list.}
\usage{
initParamList(D, Z)
}
\arguments{
\item{D}{an object of class \code{multimixSettings}---see
\code{\link{data_organise}} for details.}

\item{Z}{an \eqn{n \times q}{n by q} matrix, where \eqn{n}{n} is the number
of rows of \code{dframe} and \eqn{q}{q} is the number of components in the
mixture. During the fitting \eqn{Z_{ij}}{Zij} holds the currently estimated
probability that observation \eqn{i}{i} belongs to component \eqn{j}{j}.
Often \code{Z} is initialized to a matrix of indicator columns for a
partition of the data. It is also common to initialize \code{Z} to be the
final \code{Z} from the fitting of a simpler model.}
}
\value{
an object of class \code{multimixParamList}---see 
\code{\link{Z.to.P}} for more information.
}
\description{
Although the starting parameter list \code{P} may be specified directly, 
Note also that any matrices specified must be positive definite.
This function calculates an initial \code{P} from \code{D} and a starting
value for \code{Z}.
}
