% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.r
\name{left}
\alias{left}
\title{Map integer index N>0 back to left member of generating pair.}
\usage{
left(N)
}
\arguments{
\item{N}{positive integer scalar}
}
\value{
positive integer scalar
}
\description{
Map integer index N>0 back to left member of generating pair.
}
\examples{
left(131)
left(57)
}
