% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Z_discrete.r
\name{make_Z_discrete}
\alias{make_Z_discrete}
\title{Title}
\usage{
make_Z_discrete(d)
}
\arguments{
\item{d}{}
}
\description{
Title
}
\examples{
stage = scan(file = system.file("extdata", "Stage.txt", package = "multimix"))
stage = stage - 2
Z = make_Z_discrete(stage)
}
