% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Z_fortran.r
\name{make_Z_fortran}
\alias{make_Z_fortran}
\title{Read Z from Fortran output. Make into R matrix}
\usage{
make_Z_fortran(gr.out = "groups.out")
}
\arguments{
\item{gr.out}{}
}
\description{
Read Z from Fortran output. Make into R matrix
}
