% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Z_random.r
\name{make_Z_random}
\alias{make_Z_random}
\title{Start from random groups of similar size.}
\usage{
make_Z_random(xq, seed = 310322)
}
\arguments{
\item{xq}{}

\item{seed}{}
}
\description{
Start from random groups of similar size.
}
