% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmain2.r
\name{mmain}
\alias{mmain}
\title{Title}
\usage{
mmain(D, Z, P, eps = 1e-09)
}
\arguments{
\item{D}{an object of class \code{multimixSettings} - see
\code{\link{data_organise}} for full description.}

\item{Z}{a matrix}

\item{P}{a matrix}

\item{eps}{}

\item{nIter}{the maximum number of iteratins that multimix is allowed to run
before it terminates.}
}
\value{
an object of class \code{multimix results} which is a a list
  containing three elements: the \eqn{Z}{Z} matrix, the \eqn{P}{P} matrix,
  and a results matrix, called \code{results}, with \eqn{nIter + 2}{nIter +
  2} rows and \eqn{qq}{qq} columns.
}
\description{
Title
}
