% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStep.r
\name{mStep}
\alias{mStep}
\title{The M(aximisation) step}
\usage{
mStep(Z, D)
}
\arguments{
\item{Z}{an \eqn{n \times q}{n by q} matrix, where \eqn{n}{n} is the number
of rows of \code{dframe} and \eqn{q}{q} is the number of components in the
mixture. During the fitting \eqn{Z_{ij}}{Zij} holds the currently estimated
probability that observation \eqn{i}{i} belongs to component \eqn{j}{j}.
Commonly \code{Z} is initialized to a matrix of indicator columns for a
partition of the data.}

\item{D}{an object of class \code{multimixSettings}---see
\code{\link{data_organise}} for details.}
}
\value{
an object of class \code{multimixParamList}---see 
\code{\link{initParamList}} for more information.
}
\description{
Uses the current group membership to estimate the probabilities.
}
