% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimix-package.R
\docType{package}
\name{multimix-package}
\alias{multimix}
\alias{multimix-package}
\title{multimix: Fit Mixture Models Using the Expectation Maximisation (EM) Algorithm}
\description{
A set of functions which use the Expectation Maximisation (EM) algorithm (Dempster, A. P., Laird, N. M., and Rubin, D. B. (1977) \doi{10.1111/j.2517-6161.1977.tb01600.x} Maximum likelihood from incomplete data via the EM algorithm, Journal of the Royal Statistical Society, 39(1), 1--22) to take a finite mixture model approach to clustering. The package is designed to cluster multivariate data that have categorical and continuous variables and that possibly contain missing values. The method is described in Hunt, L. and Jorgensen, M. (1999) \doi{10.1111/1467-842X.00071} Australian & New Zealand Journal of Statistics 41(2), 153--171 and Hunt, L. and Jorgensen, M. (2003) \doi{10.1016/S0167-9473(02)00190-1} Mixture model clustering for mixed data with missing information, Computational Statistics & Data Analysis, 41(3-4), 429--440.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmcurran/multimix}
  \item Report bugs at \url{https://github.com/jmcurran/multimix/issues}
}

}
\author{
\strong{Maintainer}: James Curran \email{j.curran@auckland.ac.nz} [contributor]

Authors:
\itemize{
  \item Murray Jorgensen \email{majmurr@gmail.com}
}

}
\keyword{internal}
