% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimix.R
\name{multimix}
\alias{multimix}
\title{multimix: Find optimal model using several initiating parameters}
\usage{
multimix(
  df_long,
  fixed_pars = list(beta0_2 = 0),
  lower_bounds = lower_bounds_example,
  upper_bounds = upper_bounds_example,
  max_tries = 20,
  return_first_sucess = FALSE,
  verbose = FALSE,
  seed = 1234
)
}
\arguments{
\item{df_long}{data in long format. Columns are: Subject_ID, Time, Binary_outcome}

\item{fixed_pars}{named list of parameters to fix values (i.e. no optimization on)}

\item{lower_bounds}{named list of lower bounds to guess initial params within}

\item{upper_bounds}{named list of upper bounds to guess initial params within}

\item{max_tries}{number of initial params to be tried}

\item{return_first_sucess}{logical. If \code{TRUE} then first model that works will be returned. Otherwise will exhaust the full number of retires to find the most optimal solution}

\item{verbose}{logical. If \code{TRUE} then error messages will be displayed for each failed attempt}
}
\value{
An object of class \code{multimix_model}, which is a list containing at least:
\describe{
\item{df_long}{The original \code{df_long} data frame used for fitting.}
\item{est}{Named numeric vector of estimated parameters.}
\item{u_hat}{Estimated random effects of size \verb{[N, 2]}}
\item{logLik}{Numeric. Log-likelihood of the fitted model.}
}
The object is intended to be used with S3 methods such as \code{print()}, \code{summary()}, and \code{plot()}.
}
\description{
Due to many undefined cases of the generic function and the lack of an empirical solution
to the optimizer, this function will retry several initiating parameters
and find the most optimal model
}
