% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{myfacto}
\alias{myfacto}
\title{This file contains functions we are testing. As a consequence, none of them 
are exported, and will need the ::: operator to use them}
\usage{
myfacto(num)
}
\description{
This file contains functions we are testing. As a consequence, none of them 
are exported, and will need the ::: operator to use them
}
