% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.to.Z.r
\name{P.to.Z}
\alias{P.to.Z}
\title{Maximisation step}
\usage{
P.to.Z(P, D)
}
\arguments{
\item{D}{an object of class \code{multimixSettings}---see
\code{\link{data_organise}} for more information.}
}
\description{
Maximisation step
}
