% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.r
\name{pair.index}
\alias{pair.index}
\title{Maps integer pairs (u,v) with 0<u<v bijectively to positive integers.}
\usage{
pair.index(u, v)
}
\arguments{
\item{u}{positive integer scalar}

\item{v}{positive integer scalar}
}
\value{
integer scalar
}
\description{
Used to reduce array dimensions by replacing
       A(x,y,z) by A*(x,pair.index(y,z))
}
\examples{
pair.index(11,17)
pair.index(2,12)
}
