% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multimixResults.R
\name{plot.multimixResults}
\alias{plot.multimixResults}
\title{S3 method for plotting multimix resuls objects}
\usage{
\method{plot}{multimixResults}(x, ...)
}
\arguments{
\item{x}{an object of class \code{multimixResults} -- see
\code{\link{mmain}} for more information.}

\item{\ldots}{any other arguments to be passed to \code{plot}. Note that
because there are two calls to plot, the \ldots arguments will be passed to 
each call, and it is unlikely that this will have the desired effect.}
}
\description{
S3 method for plotting multimix resuls objects
}
\author{
James Curran
}
