% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multimixResults.R
\name{print.multimixResults}
\alias{print.multimixResults}
\title{S3 method for the printing of multimix results}
\usage{
\method{print}{multimixResults}(x, ...)
}
\arguments{
\item{x}{an object of class \code{multimixResults}---see \code{\link{mmain}}
for a description.}

\item{...}{other parameters passed to \code{print}. Not currently used.}
}
\description{
S3 method for the printing of multimix results
}
