% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.r
\name{right}
\alias{right}
\title{Map integer index N>0 back to right member of generating pair.}
\usage{
right(N)
}
\arguments{
\item{N}{positive integer scalar}
}
\value{
positive integer scalar
}
\description{
Map integer index N>0 back to right member of generating pair.
}
\examples{
right(131)
right(57)
}
\author{
Murray Jorgensen
}
