% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.store.R
\name{stat.store}
\alias{stat.store}
\title{Not currently exported as this is experimental
Murray - I have rewritten this to eliminate the use of with
and to make it clear which values are being copied from D, and which are
being redefined locally.}
\usage{
stat.store(D)
}
\description{
Not currently exported as this is experimental
Murray - I have rewritten this to eliminate the use of with
and to make it clear which values are being copied from D, and which are
being redefined locally.
}
