% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_organise.r
\name{data_organise}
\alias{data_organise}
\title{Title}
\usage{
data_organise(dframe, cdep = NULL, lcdep = NULL, minpstar = 1e-09)
}
\arguments{
\item{dframe}{}

\item{cdep}{}

\item{lcdep}{}

\item{minpstar}{Minimum increase in the log-likelihood for the EM Algorithm.
This is used as a stopping criterion. That is, if the increase in the 
log-likelihood is less than \code{minpstar} then EM Algorithm will terminate.}
}
\description{
Title
}
