% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_organise.r
\name{data_organise}
\alias{data_organise}
\title{data_organise}
\usage{
data_organise(dframe, Init_grp, cdep = NULL, lcdep = NULL,
  minpstar = 1e-09, nIt = 60)
}
\arguments{
\item{dframe}{}

\item{Init_grp}{}

\item{cdep}{}

\item{lcdep}{}

\item{minpstar}{}

\item{nIt}{}
}
\description{
data_organise
}
