% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_organise.r
\name{data_organise}
\alias{data_organise}
\title{Prepare data for use with multimix}
\usage{
data_organise(
  dframe,
  niter = 1000,
  cdep = NULL,
  lcdep = NULL,
  minpstar = 1e-09
)
}
\arguments{
\item{dframe}{a data frame containing the data set you wish to model.}

\item{niter}{the maximum number of steps to that the EM agorithm will run
before terminating.}

\item{cdep}{a list of multivariate normal cells.}

\item{lcdep}{a list of location cells.}

\item{minpstar}{Minimum increase in the log-likelihood for the EM Algorithm.
This is used as a stopping criterion. That is, if the increase in the 
log-likelihood is less than \code{minpstar} then EM Algorithm will terminate.}
}
\description{
Prepare data for use with multimix
}
\examples{
data(cancer.df)
mmObj = data_organise(cancer.df)
}
