% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Z_random.r
\name{make_Z_random}
\alias{make_Z_random}
\title{Start from random groups of similar size.}
\usage{
make_Z_random(n, xq, seed = 310322)
}
\arguments{
\item{n}{integer - the number of observations.}

\item{xq}{integer # Number of random "clusters" to be generated.}

\item{seed}{integer # Suggest using date as seed for random numbers.}
}
\value{
a matrix.
}
\description{
A large number (\eqn{n}{n}) of observations are assigned randomly into
(\eqn{xq}{xq}) clusters. It is recommended to repeat Multimix runs with a
number of different seeds to search for a log-likelihood maximum.
}
\details{
Also consider making additional clusters from observations with low
probabilities of belonging to any cluster in a previous clustering.
}
\examples{
make_Z_random(3, 271222)
}
