% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{plot.multimix_model_lite}
\alias{plot.multimix_model_lite}
\title{Plot simplified drug probability trajectories for a multi-mix model}
\usage{
\method{plot}{multi_mix_model_lite}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multi_mix_model}, typically returned by \code{fit_multi_mix_model()}.}

\item{...}{Additional arguments passed to underlying plotting functions (currently ignored).}
}
\value{
\code{ggplot} object
}
\description{
This S3 method plots the estimated probabilities of drug administration
over time for each drug class, using a lighter version of the multimix model
}
