% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multimixParamList.R
\name{print.multimixParamList}
\alias{print.multimixParamList}
\title{S3 printing method for for multimix parameter results}
\usage{
\method{print}{multimixParamList}(x, ...)
}
\arguments{
\item{x}{an object of class \code{multimixParamResults} -- see 
\code{\link{initParamList}} for more information.}

\item{...}{additional arguments passed to \code{print}.}
}
\description{
S3 printing method for for multimix parameter results
}
